requirements.txt
# Flask and web framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
Jinja2==3.1.2

# Database
PyMySQL==1.1.0
cryptography==41.0.4
SQLAlchemy==2.0.21

# Task queue
celery==5.3.2
redis==5.0.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Persian text processing and NLP
hazm==0.7.0
scikit-learn==1.3.0
numpy==1.24.3
scipy==1.11.2

# Telegram bot
python-telegram-bot==20.5

# Web server
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
markupsafe==2.1.3
werkzeug==2.3.7
itsdangerous==2.1.2

# Date and time handling
python-dateutil==2.8.2
pytz==2023.3

# Logging and monitoring
coloredlogs==15.0.1

# Development tools (optional)
flake8==6.0.0
black==23.7.0